Moved quantile functions to quantileController
