+++ nodejscomponent/lib/index.js
@@ -93,80 +93,6 @@ module.exports.getIncomeMedian = function(event, callback) {
    income / quantiles
    -------------------------------------------------------- */
 
-_getCompareIncomeQuantiles = function(queryParams, callback) {
-  var compare = queryParams.compare;
-  if(compare && compare !== "") {
-    delete queryParams[compare];
-  }
-  delete queryParams["compare"];
-
-  var sql = "SELECT QUANTILE, INCOME, " + compare + " FROM PUMS_2014_Quantiles";
-  sql = utils.appendWhereClause(sql, queryParams) + " ORDER BY QUANTILE::INT ASC;";
-
-  pg.connect(conn_options, function(err, client, done) {
-    if(err) { return callback(err); }
-
-    client.query(sql, function(err, response) {
-      done();
-      if(err) { return callback(err); }
-
-      var results = response.rows;
-      resultsObj = {};
-      _.forEach(results, function(result) {
-        if(result[compare] != "") {
-          var path = "['" + result[compare] + "']" + "['" + result["quantile"] + "%']";
-          _.set(resultsObj, path, result["income"]);
-        }
-      })
-
-      var resultsObjSorted = {};
-      Object.keys(resultsObj).sort().forEach(function(key) {
-        resultsObjSorted[key] = resultsObj[key];
-      });
-
-      return callback(err, resultsObjSorted);
-    });
-  });
-}
-
-_getIncomeQuantiles = function(queryParams, callback) {
-  var sql = "SELECT QUANTILE, INCOME FROM PUMS_2014_Quantiles";
-  sql = utils.appendWhereClause(sql, queryParams) + " ORDER BY QUANTILE::INT ASC;";
-
-  pg.connect(conn_options, function(err, client, done) {
-    if(err) { return callback(err); }
-
-    client.query(sql, function(err, response) {
-      done();
-      if(err) { return callback(err); }
-
-      var results = response.rows;
-      resultsObj = {};
-      _.forEach(results, function(result) {
-        resultsObj[result["quantile"] + "%"] = result["income"];
-      })
-
-      return callback(err, {"overall": resultsObj});
-    });
-  });
-}
-
-module.exports.getIncomeQuantiles = function(event, callback) {
-
-  var queryParams = _.pick(event, ["state", "race", "sex", "agegroup", "quantile", "compare"]);
-
-  utils.validateQueryParams(queryParams, function(err, validateCallback) {
-    if(err) { return callback(err); }
-
-    if(compare && compare !== "") {
-      _getCompareIncomeQuantiles(queryParams, callback);
-    } else {
-      delete queryParams["compare"];
-      _getIncomeQuantiles(queryParams, callback);
-    }
-  });
-
-}
 
 /* --------------------------------------------------------
    income / distribution
