+++ routes.js
@@ -2,7 +2,7 @@ var express = require('express');
 var _ = require('lodash');
 var router = express.Router();
 var distributionController = require('./controllers/distribution');
-var quantilesController = require('./controllers/quantile');
+var quantileController = require('./controllers/quantile');
 var medianController = require('./controllers/median');
 
 router.get('/distribution', distributionController.process);
